+++ CHANGELOG.md
@@ -10,7 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Unit Tests: ActiveScene
 
 ## [2.9.3] - 2022-11-29
-- Added: `HashMap.RemoveAllKeys(where)` + `HashMap.RemoveAllValues(where)`
+- Added: `HashMap.UnmapAllKeys(where)` + `HashMap.UnmapAllValues(where)`
+  - Bonus: HashMap.Enumerator now allows for deletion while iterating.
+  - Added: new HashMap unit tests
 - Added: `Transforms` static class, with space manipulation extensions.
 - Fixed: Missing constructors for SerialSet subclasses.